Project fix main (#52)
* added token

* help

* stash is now fetched before shoppinglist

* removed console.log
