+++ client/src/components/shoppingList/shoppingList.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import '../../stylesheets/shoppingList.css'
 import ShoppingListRecipe from './shoppingListRecipe';
-import StashRowElement from '../stashRowElement';
+import IngredientElement from './ingredientElement';
 
 //This is a React class it extends a React component which 
 //means that you can use all the code from the React component and it runs the
@@ -26,23 +26,6 @@ class ShoppingList extends React.Component {
 
   // Base function in react, called immediately after a component is mounted. Triggered after re-rendering
   componentDidMount() {
-    // Retrieves shoppinglist information
-    fetch(`/shoppingList`, {
-      headers: {
-        'Content-Type': 'application/json',
-        'Accept': 'application/json'
-      }
-    })
-      .then(res => res.json())
-      .then((res) => {
-        let data = {
-          shoppingListRecipes: res,
-        };
-        this.setState(data);
-      }).catch(err => {
-        console.error(err);
-      });
-
     // Retrieves mystash information
     fetch(`/stash/get`, {
       headers: {
@@ -54,6 +37,28 @@ class ShoppingList extends React.Component {
       .then((res) => {
         let data = {
           myStashIngredients: res
+        };
+        this.setState(data);
+      }).catch(err => {
+        console.error(err);
+      }).then(() => {
+        // Retrieves shoppinglist information
+        this.fetchShoppingList();
+      });
+
+  }
+
+  fetchShoppingList() {
+    fetch(`/shoppingList`, {
+      headers: {
+        'Content-Type': 'application/json',
+        'Accept': 'application/json'
+      }
+    })
+      .then(res => res.json())
+      .then((res) => {
+        let data = {
+          shoppingListRecipes: res,
         };
         this.setState(data);
       }).catch(err => {
@@ -355,17 +360,24 @@ class ShoppingList extends React.Component {
   ingredientInStash(shoppingListIngredient, ingredientIndex) {
     let isInStash = false;
     let myStashIngredients = this.state.myStashIngredients;
+    console.log(myStashIngredients)
     if (ingredientIndex === undefined) {
       return isInStash
     }
 
     myStashIngredients.forEach((ingredient, i) => {
-      if (Number(ingredient.prod_id) === Number(shoppingListIngredient.prod_id)) {
+      if (Number(ingredient.prod_id) === Number(shoppingListIngredient.props.ingredient.prod_id)) {
         isInStash = true
+        shoppingListIngredient.setState({
+          hide: true,
+        }, () => {
+          this.updateRecipePrices();
+          return isInStash
+        })
       }
     });
-
     return isInStash
+
   }
 
   removeRecipe(recipe) {
@@ -413,7 +425,13 @@ class ShoppingList extends React.Component {
     myStashComponents.push(stashRowElementInstance);
     this.setState({
       myStashComponents: myStashComponents
-    })
+    }
+      // , () => {
+      // if (this.state.myStashComponents.length)
+      // }
+    )
+
+
   }
 
   //This is the render function. This is where the
@@ -463,7 +481,7 @@ class ShoppingList extends React.Component {
                   {
                     this.state.myStashIngredients.map((ingredient) => {
                       return (
-                        <StashRowElement
+                        <IngredientElement
                           matchIngredient={(stashIngredient, subtract, matchIngredient) => this.matchIngredient(stashIngredient, subtract, matchIngredient)}
                           key={this.state.myStashIngredients.indexOf(ingredient)}
                           ingredient={ingredient} myStash={true}
