+++ client/src/components/stashRowElement.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React from "react";
 
 //This is a React class it extends a React component which
 //means that you can use all the code from the React component and it runs the
@@ -12,186 +12,48 @@ class StashRowElement extends React.Component {
     //runs before our code in the contructor
     super();
     //Your code here
-
-    this.state = {
-      hide: false,
-      boxChecked: true,
-      inited: false,
-      priceWasAdded: false,
-      wasTrashed: false,
-    };
-  }
-
-  componentDidMount() {
-    if (this.props.hasOwnProperty('shoppingList') && !this.state.inited) {
-      // TODO PROBABLY CALL THIS function with ()
-      // this.props.ingredientInStash(recipeIngredientComponent, recipeIngredientComponent.props.ingredientIndex);
-      // WHICH COULD BE GIVEN AS PROP FROM shoppingListRecipe
-      this.initShoppingListElement(this.props.isHidden);
-    }
-
-    if (this.props.hasOwnProperty('passToStashComponents') && !this.state.inited) {
-      this.initStashElement(false);
-    }
   }
 
   //Functions go here
-  initShoppingListElement(hide) {
-    if (hide) {
-      this.setState({
-        inited: true,
-        hide: hide,
-      }, () => {
-        this.props.trackShoppingListElement(this, hide)
-      });
-    }
-    else {
-      if (!this.state.inited) {
-        this.setState({
-          inited: true,
-          hide: hide,
-        }, () => {
-          this.props.trackShoppingListElement(this, hide)
-        });
-      }
-    }
-  }
 
-  initStashElement(hide) {
-    this.setState({
-      inited: true,
-      hide: hide,
-    }, () => {
-      this.props.trackStashElement(this, hide)
-    })
-  }
-
-
-
-  hideStashRowElement(stashRowElement, endPoint) {
-    this.setState({
-      wasTrashed: true,
-      boxChecked: false,
-      hide: true,
-    }, () => {
-
-    })
-
-    let params = {
-      endPoint: endPoint
-    };
-
-    console.log("")
-    console.log("pushed Trashcan")
-    if (this.props.hasOwnProperty('recipeID')) {
-      this.setState({
-        hide: !this.state.hide,
-        wasTrashed: true
-      }, () => {
-        params['recipeID'] = this.props.recipeID;
-        console.log("removing recipeIngredient")
-        this.props.removeIngredient(stashRowElement, params)
-        this.props.updateRecipePrice()
-      })
-    }
-    else {
-      params['recipeID'] = false;
-      console.log("removing stashIngredient")
-      this.props.matchIngredient(this, false, true);
-      this.props.removeIngredient(stashRowElement, params);
-    }
-  }
-
-  addItemToStash(evt) {
-    this.setState((prevState) => ({
-      boxChecked: false,
-      hide: true,
-    }), () => {
-      console.log("")
-      console.log("Adding recipeIngredient to Stash")
-      let stashRowElement = this;
-      fetch(`/stash/add`, {
-        method: 'POST',
-        headers: {
-          'Content-Type': 'application/json',
-          'Accept': 'application/json'
-        },
-        body: JSON.stringify(this.props.ingredient)
-      }).then(stashRowElement.props.updateMyStashIngredients(stashRowElement))
-        .then(this.props.matchIngredient(this, true, false, true))
+  removeIngredient() {
+    fetch(`/stash/remove/${this.props.product.prod_id}`, {
+      method: "DELETE",
+      headers: {
+        "Content-Type": "application/json",
+        Accept: "application/json",
+      },
+    }).catch((err) => {
+      console.error(err);
     });
+    this.props.updateFunction(this.props.product.prod_id);
   }
-
-  checkCheckBox(evt) {
-    this.setState((prevState) => ({
-      boxChecked: !prevState.boxChecked
-    }), () => {
-      if (this.state.boxChecked) {
-        console.log("")
-        console.log("BoxChecked")
-        this.props.matchIngredient(this, true, false);
-      }
-      else {
-        console.log("")
-        console.log("boxUnchecked")
-        this.props.matchIngredient(this, false, false);
-      }
-    });
-
-  }
+  /* this.state.product.prod_ID */
 
   //This is the render function. This is where the
   //html is.
   render() {
-    // TODO FIX BUG WHERE LOADING THE SHOPPING LIST WITH INGREDIENTS IN MY STASH DOES NOT HIDE THE INGREDIENTS IN THE RECIPES.
-    if (this.state.hide || this.state.wasTrashed) {
-      return null;
-    }
-
-    if (this.props.hasOwnProperty('shoppingList')) {
-      if (!this.state.priceWasAdded) {
-        this.setState({
-          priceWasAdded: true
-        })
-      }
-      return (
-        <tr>
-          <td>{this.props.ingredient ? this.props.ingredient.title : ""}</td>
-          <td className="right-align">{this.props.ingredient ? this.props.ingredient.price : ""} kr.</td>
-          <td className="right-align">
-            <button type="button" onClick={() => { this.hideStashRowElement(this.props.ingredient, 'removeIngredientFromShoppingList/') }}>
-              <i className="fa fa-trash"></i></button>
-          </td>
-          <td className="right-align center" width="2%">
-            <div className="form-check align-middle">
-              <input className="form-check-input" type="checkbox" onChange={(evt) => this.addItemToStash(evt)}
-                id="flexCheckChecked" checked={!this.state.boxChecked}>
-              </input>
-            </div>
-          </td>
-        </tr>
-      );
-    }
-    else if (this.props.hasOwnProperty('myStash')) {
-      return (
-        <tr>
-          <td>{this.props.ingredient ? this.props.ingredient.title : ""}</td>
-          <td className="right-align">{this.props.ingredient ? this.props.ingredient.amount : ""} {this.props.ingredient ? this.props.ingredient.unit : ""}</td>
-          <td className="right-align">
-            <button type="button" onClick={() => { this.hideStashRowElement(this.props.ingredient, '/stash/remove/') }}>
-              <i className="fa fa-trash"></i>
-            </button>
-          </td>
-          <td className="right-align center" width="2%">
-            <div className="form-check align-middle">
-              <input className="form-check-input" type="checkbox" onChange={(evt) => this.checkCheckBox(evt)}
-                id="flexCheckChecked" checked={this.state.boxChecked}>
-              </input>
-            </div>
-          </td>
-        </tr>
-      )
-    }
+    return (
+      <tr class="table-content  table-rounded">
+        <td>{this.props.product ? this.props.product.title : ""}</td>
+        <td class="table-content-secondary">
+          {this.props.product
+            ? this.props.product.amount + " " + this.props.product.unit
+            : ""}
+        </td>
+        <td class="right-align-text">
+          <button
+            type="button"
+            class="deleteButton"
+            onClick={() => {
+              this.removeIngredient();
+            }}
+          >
+            <i class="fa fa-trash"></i>
+          </button>
+        </td>
+      </tr>
+    );
   }
 }
